update to respond to post requests
